renaming stepper -> step-main
