+++ mqtt-mapping-ui/mqtt-mapping/src/mqtt-mapping/step-main/stepper-model.ts
@@ -19,7 +19,7 @@
  * @authors Christof Strack
  */
 
-import { Direction } from "./../../shared/mapping.model";
+import { Direction } from "../../shared/mapping.model";
 
 export enum EditorMode {
   CREATE = "CREATE",
