+++ mqtt-mapping-ui/mqtt-mapping/src/mqtt-mapping/step-main/substitution/substitution-validation.pipe.ts
