+++ mqtt-mapping-ui/mqtt-mapping/src/mqtt-mapping/step-main/util.ts
